Use 'git.fetch_tag()' or 'git.fetch_branch()' instead of 'git.clone()'
'git clone' downloads all the branches, so in a case that the repository of a module
that is about to be cloned, contains also branches with compiled binary versions,
these branches will be downloaded as well, resulting in higher bandwidth usage.

Thus we clone only the desired tag/branch using:
  1. 'git init'
  2. -- 'git fetch -u REPO +refs/tags/TAG:refs/tags/TAG'
     or 'git fetch -u REPO +refs/heads/BRANCH:BRANCH'
  3. 'git checkout TAG' or 'git checkout BRANCH'

In particular, when downloading the 'scm' version of some module, we fetch its
'master' branch. Otherwise we fetch the tag matching the desired module version.